[test] Use tuple instead of list to index underlying raster data.
This adresses a FutureWarning raised by numpy:

FutureWarning: Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will result either in an error or a different result.
  expected_raster_values = np.tile(raster_data[data_array_indices], 2)